requirements.txt
Flask==0.10.1
azure-common==1.1.4 
azure-nspkg==1.0.0 
azure-storage==0.33.0
